+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
 - Fix CUDA `sparse_to_dense`. ([#386])
+- Improve pseudo inverse spectrum solver. ([#388])
 
 ## [v0.25.2]
 Release date: 2025-02-02
@@ -107,3 +108,4 @@ Release date: 2024-11-13
 [#380]: https://github.com/qutip/QuantumToolbox.jl/issues/380
 [#383]: https://github.com/qutip/QuantumToolbox.jl/issues/383
 [#386]: https://github.com/qutip/QuantumToolbox.jl/issues/386
+[#388]: https://github.com/qutip/QuantumToolbox.jl/issues/388